## Summary

Includes all current workspace changes as requested, plus the
mobile/server troubleshooting fixes from this session.

## Included in this PR

- All currently modified, added, and deleted files in the working tree
- Mobile connectivity diagnostics improvements:
  - Better server test endpoint and error visibility
  - Better login network error surfacing
- Frontend import diagnostics improvements:
- Better import error extraction for non-JSON proxy responses (e.g. 413
HTML pages)
  - Regression tests around import error messaging
- Mobile Docker build script hardening for constrained hosts:
  - Lower-memory Gradle defaults
  - Serialized workers
  - Isolated Gradle user home to avoid cache/journal lock contention
  - APK copy logic for release/debug artifacts

## Notes

- This is a broad PR by request ("include all").
- Mobile build on this host still showed OOM kills (exit 137) in
containerized Gradle execution, indicating host memory pressure beyond
script-level tuning.

+++ backend/src/Controller/PartController.php
@@ -362,7 +362,7 @@ class PartController extends AbstractController
             $part->setPrice($data['price']);
         }
         if (isset($data['quantity'])) {
-            $part->setQuantity((int)$data['quantity']);
+            $part->setQuantity($data['quantity']);
         }
         if (isset($data['cost'])) {
             // Ensure entity expects a string - cast numeric values to string
@@ -372,9 +372,9 @@ class PartController extends AbstractController
         // If price/quantity provided (or both exist on entity), compute cost
         if (array_key_exists('price', $data) || array_key_exists('quantity', $data)) {
             $p = $data['price'] ?? $part->getPrice();
-            $q = array_key_exists('quantity', $data) ? (int)$data['quantity'] : $part->getQuantity();
+            $q = array_key_exists('quantity', $data) ? (float)$data['quantity'] : (float)$part->getQuantity();
             if ($p !== null && $q !== null) {
-                $computed = number_format(((float)$p) * ((int)$q), 2, '.', '');
+                $computed = number_format(((float)$p) * $q, 2, '.', '');
                 $part->setCost((string)$computed);
             }
         }
